+++ Tests/Integration/Comments.Integration.Tests.ps1
@@ -14,7 +14,7 @@ BeforeDiscovery {
 }
 
 InModuleScope JiraPS {
-    Describe "Issue Comments" -Tag 'Integration' -Skip:$Skip {
+    Describe "Issue Comments" -Tag 'Integration', 'Server', 'Cloud' -Skip:$Skip {
         BeforeAll {
             . "$PSScriptRoot/../Helpers/IntegrationTestTools.ps1"