+++ Tests/Integration/New-JiraIssue.Integration.Tests.ps1
@@ -14,7 +14,7 @@ BeforeDiscovery {
 }
 
 InModuleScope JiraPS {
-    Describe "New-JiraIssue" -Tag 'Integration', 'Smoke' -Skip:$Skip {
+    Describe "New-JiraIssue" -Tag 'Integration', 'Smoke', 'Server', 'Cloud' -Skip:$Skip {
         BeforeAll {
             . "$PSScriptRoot/../Helpers/IntegrationTestTools.ps1"